style: move identical windows & linux code to shared file
