+++ npm/metrics/prometheus-values.go
@@ -9,6 +9,30 @@ import (
 
 var errNotCollector = fmt.Errorf("error: summary metric is not a collector")
 
+func counterValue(counter prometheus.Counter) (int, error) {
+	dtoMetric, err := getDTOMetric(counter)
+	if err != nil {
+		return 0, err
+	}
+	return int(dtoMetric.Counter.GetValue()), nil
+}
+
+func histogramVecCount(histogramVec *prometheus.HistogramVec, labels prometheus.Labels) (int, error) {
+	collector, ok := histogramVec.With(labels).(prometheus.Collector)
+	if !ok {
+		return 0, errNotCollector
+	}
+	return histogramCount(collector)
+}
+
+func histogramCount(histogram prometheus.Collector) (int, error) {
+	dtoMetric, err := getDTOMetric(histogram)
+	if err != nil {
+		return 0, err
+	}
+	return int(dtoMetric.Histogram.GetSampleCount()), nil
+}
+
 // getValue returns a Gauge metric's value.
 // This function is slow.
 func getValue(gaugeMetric prometheus.Gauge) (int, error) {
